# Refs #27624 -- Changed Query.explain_info to namedtuple.