* Turn explain info into one attribute instead of three
* Remove unnecessary `if ... is None` paths, which are already shallow-copied in the `__dict__.copy()` call.

ticket-27624