@kezabelle 2021-09-15
I've got a separate branch with much the same changes (only setting `if not None`), having noted basically the same. However I never bench'd the difference in isolation (it's part of a bigger changeset) - can we do so?

As you mentioned on the ticket:

> But it may not be worth the effort to change and profile.

Honestly I think any and every effort to speed up (or obviate) cloning costs is worthwhile, as it is a large source of slowdown in things like `prefetch_related`. I've got a change which could land subsequent to this which may help, and a couple of other ideas off the back of that.