@adamchainz 2021-09-15
> I've got a separate branch with much the same changes (only setting `if not None`), having noted basically the same.

Link? Did you make a PR?

> However I never bench'd the difference in isolation (it's part of a bigger changeset) - can we do so?

Basic benchmark on Python 3.9.7.

Before:

```
In [1]: from django.db.models.sql.query import Query

In [2]: %timeit Query('somemodel').clone()
7.09 µs ± 53.9 ns per loop (mean ± std. dev. of 7 runs, 100000 loops each)
```

After:

```
In [1]: from django.db.models.sql.query import Query

In [2]: %timeit Query('somemodel').clone()
6.87 µs ± 106 ns per loop (mean ± std. dev. of 7 runs, 100000 loops each)
```

> Honestly I think any and every effort to speed up (or obviate) cloning costs is worthwhile, as it is a large source of slowdown in things like `prefetch_related`. I've got a change which could land subsequent to this which may help, and a couple of other ideas off the back of that.

Okay, if we're happy to carry on stacking micro-optimizations then I'm happy to see them coming!