@kezabelle 2021-09-15
> Link? Did you make a PR?

Apologies if it came off like you were treading on my toes (if anything, I'm treading on yours), there's no public branch or PR; it was part of a larger changeset which I was considering proposing (which is _also_ something you were looking at in the same ticket) as a ticket once I've found time to sit down and step through the various commits noting any differences/regressions.

There are a couple of "related" PRs (insofar as my target is "prefetch_related is slow, lemme try and fix that") floating around on here, but none that directly touch this bit.